refactor: extract browser interaction helpers
